fix(front): comment out improper nest css
Upon investigation, it was observed that the .searchContainer in
rate.module.css is not utilized anywhere in the project.
This unused code segment was causing errors, manifesting as a
slowdown in compile speed and imposing a heavy workload on my computer.

The error message is
Nested CSS was detected, but CSS nesting has not been configured correctly.
Please enable a CSS nesting plugin *before* Tailwind in your configuration.
See how here: https://tailwindcss.com/docs/using-with-preprocessors#nesting

Following this change, the web application now functions smoothly without
triggering a full-page reload. However, attempting to reload the rate or
database pages may still result in three errors (Hydration failed) due to
discrepancies between the initial UI and the server-rendered content.

The error messages are
Error: Hydration failed because the initial UI does not match what was
rendered on the server.
See more info here: https://nextjs.org/docs/messages/react-hydration-error
Error: There was an error while hydrating. Because the error happened
outside of a Suspense boundary, the entire root will switch to client rendering.